moved tag-based formatter file
