+++ packages/deprecation-crawler/src/lib/output-formatters/index.ts
@@ -1,6 +1,6 @@
 export {addCommentToRepository} from './add-comments-to-repository';
 export {generateRawJson} from './generate-raw-json';
-import {generateTagBasedFormatter} from './generate-markdown';
+import {generateTagBasedFormatter} from './markdown/generate-markdown';
 import {generateGroupBasedFormatter} from './markdown/group-based.md.formatter';
 
 export const builtInFormatter = {
